# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AquaForge API",
    "description": "Swim Meet Optimization API - Backend services for AquaForge.ai",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/optimize": {
      "post": {
        "tags": [
          "Optimization (Legacy)"
        ],
        "summary": "Optimize Lineup",
        "description": "Run lineup optimization for a swim meet.\n\nThis endpoint accepts team data and optimization parameters,\nthen returns the optimal lineup assignments.\n\nArgs:\n    request: Optimization request with team data and parameters\n\nReturns:\n    Optimization results with scores and lineup assignments",
        "operationId": "optimize_lineup_api_v1_optimize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptimizationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimizationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/data/upload": {
      "post": {
        "tags": [
          "Data"
        ],
        "summary": "Upload Team File",
        "description": "Upload and process a team roster file.\n\nSupports Excel (.xlsx, .xls), CSV, and JSON files.\n\nArgs:\n    file: The roster file to upload\n    team_type: Either 'seton' or 'opponent'\n    team_name: Optional team name override\n\nReturns:\n    Processing results with swimmer and entry counts",
        "operationId": "upload_team_file_api_v1_data_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_team_file_api_v1_data_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamDataResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/data/team": {
      "post": {
        "tags": [
          "Data"
        ],
        "summary": "Submit Team Data",
        "description": "Submit team data directly as JSON.\n\nAlternative to file upload for programmatic access.",
        "operationId": "submit_team_data_api_v1_data_team_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamDataResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/data/templates/{template_type}": {
      "get": {
        "tags": [
          "Data"
        ],
        "summary": "Get Data Template",
        "description": "Get a data template for file uploads.\n\nArgs:\n    template_type: Either 'csv' or 'json'",
        "operationId": "get_data_template_api_v1_data_templates__template_type__get",
        "parameters": [
          {
            "name": "template_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Template Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/dual-meet/optimize": {
      "post": {
        "tags": [
          "Dual Meet",
          "Dual Meet"
        ],
        "summary": "Optimize dual meet lineup",
        "description": "Optimize lineup for a head-to-head dual meet.\n    \n    Takes rosters for both teams and returns the optimal lineup\n    with scoring predictions and recommendations.\n    \n    **Optimization Methods:**\n    - `gurobi`: Integer programming (most accurate, requires Gurobi license)\n    - `nash`: Nash equilibrium game theory\n    - `heuristic`: Fast greedy algorithm\n    \n    **Scoring:**\n    - Individual: Top 7 score [8, 6, 5, 4, 3, 2, 1]\n    - Relays: Top 3 score [10, 5, 3]\n    - Total: 232 points",
        "operationId": "optimize_dual_meet_api_v2_dual_meet_optimize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DualMeetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DualMeetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/dual-meet/validate": {
      "post": {
        "tags": [
          "Dual Meet",
          "Dual Meet"
        ],
        "summary": "Validate dual meet data",
        "description": "Validate team rosters without running full optimization.",
        "operationId": "validate_dual_meet_api_v2_dual_meet_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DualMeetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Validate Dual Meet Api V2 Dual Meet Validate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_team_file_api_v1_data_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "team_type": {
            "type": "string",
            "title": "Team Type"
          },
          "team_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Name"
          }
        },
        "type": "object",
        "required": [
          "file",
          "team_type"
        ],
        "title": "Body_upload_team_file_api_v1_data_upload_post"
      },
...
      "DualMeetRequest": {
        "properties": {
          "our_team": {
            "$ref": "#/components/schemas/TeamRosterModel",
            "description": "Our team roster"
          },
          "opponent": {
            "$ref": "#/components/schemas/TeamRosterModel",
            "description": "Opponent team roster"
          },
          "options": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DualMeetOptions"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optimization options"
          }
        },
        "type": "object",
        "required": [
          "our_team",
          "opponent"
        ],
        "title": "DualMeetRequest",
        "description": "Request for dual meet optimization.",
        "example": {
          "opponent": {
            "entries": [
              {
                "event": "100 Free",
                "grade": 12,
                "swimmer": "Bob Jones",
                "time": 53.01
              }
            ],
            "team_name": "Trinity"
          },
          "options": {
            "enforce_fatigue": false,
            "method": "gurobi"
          },
          "our_team": {
            "entries": [
              {
                "event": "100 Free",
                "grade": 11,
                "swimmer": "John Smith",
                "time": 52.34
              },
              {
                "event": "50 Free",
                "grade": 10,
                "swimmer": "Jane Doe",
                "time": 26.12
              }
            ],
            "team_name": "Seton"
          }
        }
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "Standard error response."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OptimizationRequest": {
        "properties": {
          "seton_data": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Seton Data",
            "description": "Seton team data"
          },
          "opponent_data": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Opponent Data",
            "description": "Opponent team data"
          },
          "optimizer_backend": {
            "$ref": "#/components/schemas/OptimizerBackend",
            "description": "Optimization engine to use (GUROBI = accurate MILP, HEURISTIC = fast fallback)",
            "default": "gurobi"
          },
          "max_individual_events": {
            "type": "integer",
            "maximum": 6.0,
            "minimum": 1.0,
            "title": "Max Individual Events",
            "default": 4
          },
          "enforce_fatigue": {
            "type": "boolean",
            "title": "Enforce Fatigue",
            "default": true
          },
          "use_relay_optimization": {
            "type": "boolean",
            "title": "Use Relay Optimization",
            "default": true
          },
          "robust_mode": {
            "type": "boolean",
            "title": "Robust Mode",
            "description": "Run multi-scenario evaluation",
            "default": false
          },
          "scoring_type": {
            "type": "string",
            "title": "Scoring Type",
            "description": "Scoring rules (visaa_top7 or standard_top5)",
            "default": "visaa_top7"
          },
          "seton_team_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seton Team Name",
            "default": "Seton"
          },
          "opponent_team_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opponent Team Name",
            "default": "Opponent"
          }
        },
        "type": "object",
        "required": [
          "seton_data",
          "opponent_data"
        ],
        "title": "OptimizationRequest",
        "description": "Request model for optimization."
      },
      "OptimizationResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "seton_score": {
            "type": "number",
            "title": "Seton Score"
          },
          "opponent_score": {
            "type": "number",
            "title": "Opponent Score"
          },
          "score_margin": {
            "type": "number",
            "title": "Score Margin"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/OptimizationResult"
            },
            "type": "array",
            "title": "Results"
          },
          "statistics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Statistics"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "optimization_time_ms": {
            "type": "number",
            "title": "Optimization Time Ms"
          },
          "championship_standings": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Championship Standings",
            "description": "Full team standings for championship mode: [{rank, team, points}, ...]"
          },
          "event_breakdowns": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Breakdowns",
            "description": "Per-event point breakdown for all teams"
          },
          "swing_events": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Swing Events",
            "description": "Events where small improvements yield significant gains"
          },
          "analytics": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analytics",
            "description": "Advanced analytics: monte_carlo, fatigue_warnings, nash_insights, relay_tradeoffs"
          }
        },
        "type": "object",
        "required": [
          "success",
          "seton_score",
          "opponent_score",
          "score_margin",
          "results",
          "optimization_time_ms"
        ],
        "title": "OptimizationResponse",
        "description": "Response from optimization request."
      },
...
      "OptimizerBackend": {
        "type": "string",
        "enum": [
          "heuristic",
          "gurobi"
        ],
        "title": "OptimizerBackend",
        "description": "Available optimization backends."
      },
...
      "TeamDataRequest": {
        "properties": {
          "team_name": {
            "type": "string",
            "title": "Team Name",
            "description": "Team name"
          },
          "team_type": {
            "$ref": "#/components/schemas/TeamType",
            "description": "Team classification (seton/opponent)"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/SwimmerEntry"
            },
            "type": "array",
            "title": "Entries",
            "description": "List of swimmer entries"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "team_name",
          "team_type",
          "entries"
        ],
        "title": "TeamDataRequest",
        "description": "Request model for team data upload."
      },
      "TeamDataResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "team_name": {
            "type": "string",
            "title": "Team Name"
          },
          "team_type": {
            "$ref": "#/components/schemas/TeamType"
          },
          "swimmer_count": {
            "type": "integer",
            "title": "Swimmer Count"
          },
          "entry_count": {
            "type": "integer",
            "title": "Entry Count"
          },
          "events": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Events",
            "description": "Unique events"
          },
          "data": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Data",
            "description": "Parsed entry data"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "file_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Hash"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "teams": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Teams",
            "description": "Team codes for championship meets"
          }
        },
        "type": "object",
        "required": [
          "success",
          "team_name",
          "team_type",
          "swimmer_count",
          "entry_count"
        ],
        "title": "TeamDataResponse",
        "description": "Response after processing team data."
      },
...
      "TeamType": {
        "type": "string",
        "enum": [
          "seton",
          "opponent"
        ],
        "title": "TeamType",
        "description": "Team type classification."
      },
...
    }
  }
}